csapp

+++ csapp/src/memory/dram.h
@@ -1,3 +1,3 @@
 
 #define MM_LEN 1000
-uint8_t mm[MM_LEN];
+uint64_t mm[MM_LEN];
